requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.9.0
